feat(budget): open set budget as dialog
