+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.28.0] - 2025-05-19
+### Added
+- Set Budget now opens as a dialog similar to New Transaction.
+- Goals can now be linked to a specific category.
 
 ## [0.27.0] - 2025-05-19
 ### Added
