--- app/src/main/java/dev/pandesal/sbp/presentation/categories/budget/SetBudgetSheet.kt
@@ -1,189 +1,0 @@
-package dev.pandesal.sbp.presentation.categories.budget
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.ElevatedCard
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.imePadding
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.FloatingToolbarDefaults
-import androidx.compose.material3.HorizontalFloatingToolbar
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.SheetState
-import androidx.compose.material3.Text
-import androidx.compose.material3.ToggleButton
-import androidx.compose.material3.ButtonGroupDefaults
-import androidx.compose.material3.rememberDatePickerState
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.twotone.DateRange
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import java.math.BigDecimal
-import java.time.Instant
-import java.time.LocalDate
-import java.time.ZoneId
-
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-fun SetBudgetSheet(
-    initialAmount: BigDecimal = BigDecimal.ZERO,
-    onSubmit: (amount: BigDecimal, isGoal: Boolean, dueDate: LocalDate?) -> Unit,
-    onCancel: () -> Unit,
-    onDismissRequest: () -> Unit,
-    sheetState: SheetState = rememberModalBottomSheetState()
-) {
-    var amount by remember { mutableStateOf(initialAmount) }
-    var selectedTab by remember { mutableIntStateOf(0) }
-    var showDatePicker by remember { mutableStateOf(false) }
-    val datePickerState = rememberDatePickerState()
-    var selectedDate by remember { mutableStateOf<LocalDate?>(null) }
-
-    ModalBottomSheet(
-        onDismissRequest = onDismissRequest,
-        sheetState = sheetState
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(24.dp)
-                .imePadding(),
-            horizontalAlignment = Alignment.End
-        ) {
-            Spacer(modifier = Modifier.weight(1f))
-
-            ElevatedCard(
-                shape = androidx.compose.foundation.shape.RoundedCornerShape(50),
-                elevation = CardDefaults.elevatedCardElevation(defaultElevation = 16.dp)
-            ) {
-                IconButton(
-                    modifier = Modifier
-                        .height(24.dp)
-                        .padding(4.dp),
-                    onClick = {
-                        onCancel()
-                        onDismissRequest()
-                    }
-                ) {
-                    Icon(Icons.Filled.Close, contentDescription = null)
-                }
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ElevatedCard(
-                shape = androidx.compose.foundation.shape.RoundedCornerShape(10),
-                elevation = CardDefaults.elevatedCardElevation(defaultElevation = 16.dp)
-            ) {
-                Column(modifier = Modifier.padding(16.dp)) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(ButtonGroupDefaults.ConnectedSpaceBetween)
-                    ) {
-                        val options = listOf("Budget", "Goal")
-                        options.forEachIndexed { index, label ->
-                            ToggleButton(
-                                checked = selectedTab == index,
-                                onCheckedChange = { selectedTab = index },
-                                modifier = Modifier.weight(1f),
-                                shapes = when (index) {
-                                    0 -> ButtonGroupDefaults.connectedLeadingButtonShapes()
-                                    options.lastIndex -> ButtonGroupDefaults.connectedTrailingButtonShapes()
-                                    else -> ButtonGroupDefaults.connectedMiddleButtonShapes()
-                                }
-                            ) {
-                                Text(label, color = if (selectedTab == index) Color.White else Color.Black)
-                            }
-                        }
-                    }
-
-                    OutlinedTextField(
-                        value = amount.toString(),
-                        onValueChange = { amount = it.toBigDecimalOrNull() ?: BigDecimal.ZERO },
-                        label = { Text("Amount") },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 16.dp),
-                        singleLine = true
-                    )
-
-                    if (selectedTab == 1) {
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(top = 16.dp)
-                                .clickable { showDatePicker = true },
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text(selectedDate?.toString() ?: "Select Due Date")
-                            Icon(Icons.TwoTone.DateRange, contentDescription = null)
-                        }
-                    }
-                }
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            HorizontalFloatingToolbar(
-                modifier = Modifier.align(Alignment.CenterHorizontally),
-                expanded = true,
-                floatingActionButton = {
-                    FloatingToolbarDefaults.VibrantFloatingActionButton(
-                        onClick = {
-                            onSubmit(amount, selectedTab == 1, selectedDate.takeIf { selectedTab == 1 })
-                            onDismissRequest()
-                        },
-//                        enabled = amount > BigDecimal.ZERO
-                    ) {
-                        Icon(Icons.Default.Check, contentDescription = null)
-                    }
-                },
-                content = {}
-            )
-
-            if (showDatePicker) {
-                DatePickerDialog(
-                    onDismissRequest = { showDatePicker = false },
-                    confirmButton = {
-                        IconButton(onClick = {
-                            datePickerState.selectedDateMillis?.let { millis ->
-                                selectedDate = Instant.ofEpochMilli(millis).atZone(ZoneId.systemDefault()).toLocalDate()
-                            }
-                            showDatePicker = false
-                        }) { Icon(Icons.Default.Check, contentDescription = null) }
-                    },
-                    dismissButton = {
-                        IconButton(onClick = { showDatePicker = false }) { Icon(Icons.Default.Close, contentDescription = null) }
-                    }
-                ) {
-                    DatePicker(state = datePickerState)
-                }
-            }
-        }
-    }
-}
